+++ service/controller/azureconfig/handler/volumebindingmigration/resource.go
@@ -5,6 +5,7 @@ import (
 	"context"
 
 	providerv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/provider/v1alpha1"
+	"github.com/giantswarm/errors/tenant"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
@@ -65,7 +66,15 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	var tenantClusterK8sClient ctrl.Client
 	{
 		tenantClusterK8sClient, err = r.getTenantClusterClient(ctx, &cr)
-		if err != nil {
+		if tenant.IsAPINotAvailable(err) || tenantcluster.IsTimeout(err) {
+			// The kubernetes API is not reachable. This usually happens when a new cluster is being created.
+			// This makes the whole controller to fail and stops next handlers from being executed even if they are
+			// safe to run. We don't want that to happen so we just return and we'll try again during next loop.
+			r.logger.Debugf(ctx, "tenant API not available yet")
+			r.logger.Debugf(ctx, "canceling resource")
+
+			return nil
+		} else if err != nil {
 			return microerror.Mask(err)
 		}
 	}