+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Avoid blocking the whole `AzureConfig` handler on cluster creation because we can't update the `StorageClasses`.
+
 ## [5.4.0] - 2021-02-05
 
 ### Changed